WIP: mavlink-adapter
Adding files of the current state for the mavlink-adapter. It is way apart from being completed in this step just the basic structure was added but is not completed yet.
In detail: This Mavlink-Adapter aims to combine the already existing dronefleet/mavlink with the vert.x bus so that both systems can be used simultaneously .
The Basic Message Structure is already implemented though the parser into raw-messages still needs some work. After this the Listener and the Sender also still need to be written.