Updating project structure and enhancing code
This patch enhances lots of the already written classes by improving the code handling different RawMavlink Messages. This means that the code shrank while on the other side new classes were added such as a SecretKeyStore for handling the password securely and a InvalidChecksumException which will be thrown if message is wrongly signed -> bad packets!