Adding Java-Docs and fixing some obvious errors
Adding Java-Docs to all classes (might need improvement (especially MavlinkParser, Transmitter and Receiver).
On top of this a lot of obvious mistakes were eliminated.
xml2record-tool was updated to v1.3.7.
Deleting unnecessary classes which were artifacts of the development process.

This is still WIP and should not be merged due to missing tests.