Proposing changes to core/connection module
The connection module is less than ideal especially regarding the consistency.

This proposal introduces a consistent way of sending received connection data via the bus to all verticles which need it. As all incoming connection-data at least consists of a byte-stream and some sender information, this can be used to create one universal ConnectionData-class and let the interfaces create their own participants (e.g. TcpConn.Participant) which implement ConnectionDetails.
ConnectionType then helps to provide the parsing-class for the participants, which can be achieved via reflection.